QOL Improvement Recent Files
Adds recent files.  Makes installer behave a little better and cleans up registry.  Trying to get it to clean up the filesystem objects as well but that'll likely require more work.  Also VS Installer Projects are still broken, it does NOT remove the old version from the Add/Remove Programs during an upgrade (despite all the documentation and claims to the contrary)